Move all diff editing logic into the tool
