+++ extensions/positron-assistant/src/tools/notebookUtils.ts
@@ -9,7 +9,6 @@ import * as xml from '../xml.js';
 import { calculateSlidingWindow, filterNotebookContext, MAX_CELLS_FOR_ALL_CELLS_CONTEXT, getOriginalContentLength } from '../notebookContextFilter.js';
 import { isRuntimeSessionReference } from '../utils.js';
 import { log } from '../extension.js';
-import { NotebookCellEditProposal, NotebookCellEditType, NotebookCellEditOperation, PositronAssistantToolName } from '../types.js';
 
 /**
  * Maximum preview length per cell for confirmations (characters)
@@ -716,146 +715,4 @@ export function serializeNotebookContextAsUserMessage(
 	parts.push('</notebook-state>');
 
 	return parts.join('\n');
-}
-
-// --- Notebook Edit Proposal Handling ---
-
-/**
- * Parses a notebook edit proposal from a tool result.
- * Returns the proposal if it's a valid notebook cell edit, null otherwise.
- */
-function parseNotebookEditProposal(result: vscode.LanguageModelToolResult): NotebookCellEditProposal | null {
-	// Check if the result contains a JSON edit proposal
-	if (result.content.length === 0) {
-		return null;
-	}
-
-	const firstPart = result.content[0];
-	if (!(firstPart instanceof vscode.LanguageModelTextPart)) {
-		return null;
-	}
-
-	try {
-		const proposal = JSON.parse(firstPart.value);
-		if (proposal?.type === NotebookCellEditType && proposal.operation === NotebookCellEditOperation) {
-			return proposal as NotebookCellEditProposal;
-		}
-	} catch {
-		// Not a JSON proposal, return null
-	}
-
-	return null;
-}
-
-/**
- * Applies a notebook edit proposal by converting it to a text edit.
- * This routes the edit through the chat editing system for diff views.
- */
-async function applyNotebookEditProposal(
-	proposal: NotebookCellEditProposal,
-	response: vscode.ChatResponseStream
-): Promise<void> {
-	// Validate that we have either cellUri or notebookUri
-	if (!proposal.cellUri && !proposal.notebookUri) {
-		throw new Error('Either cellUri or notebookUri must be provided in edit proposal');
-	}
-
-	// Prefer cellUri if available, otherwise resolve from notebook document
-	let cellDocUri: vscode.Uri;
-	if (proposal.cellUri) {
-		try {
-			cellDocUri = vscode.Uri.parse(proposal.cellUri);
-		} catch (error) {
-			const errorMsg = error instanceof Error ? error.message : String(error);
-			throw new Error(`Invalid cell URI format: ${proposal.cellUri}. ${errorMsg}`);
-		}
-	} else {
-		// Fallback: open notebook document and get cell URI by index
-		try {
-			const notebookUri = vscode.Uri.parse(proposal.notebookUri);
-			const notebookDoc = await vscode.workspace.openNotebookDocument(notebookUri);
-
-			// Validate cellIndex is a valid integer
-			if (!Number.isInteger(proposal.cellIndex) || proposal.cellIndex < 0) {
-				throw new Error(`Invalid cellIndex: ${proposal.cellIndex}. Must be a non-negative integer.`);
-			}
-
-			// cellAt() auto-adjusts indices per VS Code API, but validate bounds explicitly
-			if (proposal.cellIndex >= notebookDoc.cellCount) {
-				throw new Error(`Cell index ${proposal.cellIndex} is out of bounds (notebook has ${notebookDoc.cellCount} cells)`);
-			}
-
-			const cell = notebookDoc.cellAt(proposal.cellIndex);
-			cellDocUri = cell.document.uri;
-		} catch (error) {
-			// Re-throw with context if this is an error we didn't already throw
-			if (error instanceof Error && !error.message.includes('Invalid cellIndex') && !error.message.includes('out of bounds')) {
-				throw new Error(`Failed to open notebook document: ${error.message}`);
-			}
-			throw error;
-		}
-	}
-
-	// Get the current cell content to compute a proper text edit
-	const cellDoc = await vscode.workspace.openTextDocument(cellDocUri);
-	const currentContent = cellDoc.getText();
-	const newContent = proposal.newContent;
-
-	// Only create edit if content actually changed
-	if (currentContent === newContent) {
-		return;
-	}
-
-	// Create a full document replacement edit
-	// This will be diffed by the chat editing system
-	const edit = new vscode.TextEdit(
-		new vscode.Range(0, 0, cellDoc.lineCount, 0),
-		newContent
-	);
-
-	// Emit the text edit through the chat editing system
-	response.textEdit(cellDocUri, edit);
-}
-
-/**
- * Handles notebook tool results that contain edit proposals.
- * If the result is a notebook cell edit proposal, applies it via response.textEdit()
- * and returns a result message. Otherwise, returns undefined to indicate the result
- * should be passed through unchanged.
- *
- * @param toolName The name of the tool that was invoked
- * @param result The tool result to check
- * @param response The chat response stream for emitting edits
- * @returns The result to use if handled, or undefined if not handled
- */
-export async function handleNotebookEditProposal(
-	toolName: string,
-	result: vscode.LanguageModelToolResult,
-	response: vscode.ChatResponseStream
-): Promise<vscode.LanguageModelToolResult | undefined> {
-	// Only handle EditNotebookCells tool
-	if (toolName !== PositronAssistantToolName.EditNotebookCells) {
-		return undefined;
-	}
-
-	const proposal = parseNotebookEditProposal(result);
-	if (!proposal) {
-		return undefined;
-	}
-
-	try {
-		await applyNotebookEditProposal(proposal, response);
-		// Return a success message instead of the raw proposal
-		return new vscode.LanguageModelToolResult([
-			new vscode.LanguageModelTextPart(`Successfully proposed edit to cell ${proposal.cellIndex}`)
-		]);
-	} catch (error) {
-		const errorMessage = error instanceof Error ? error.message : String(error);
-		log.error(`[tool] Failed to apply notebook edit proposal: ${errorMessage}`);
-		// Return the error as a tool result so the model can handle it gracefully
-		return new vscode.LanguageModelToolResult([
-			new vscode.LanguageModelTextPart(`Error applying notebook cell edit to cell ${proposal.cellIndex}`),
-			new vscode.LanguageModelTextPart(errorMessage)
-		]);
-	}
 }
